requirements.txt
beautifulsoup4
black
coverage
flake8
gitpython
html5validator
isort
ivy
pandas
pillow
plotly
pwlf
pybtex
pydriller
pymdown-extensions
pytest
requests
snakemake
xlrd
yaml-header-tools
